Flatten caf::opencl::detail namespace
