+++ libcaf_opencl/caf/opencl/arguments.hpp
@@ -26,11 +26,8 @@
 #include "caf/optional.hpp"
 
 #include "caf/opencl/mem_ref.hpp"
-#include "caf/opencl/detail/core.hpp"
 
 namespace caf {
-namespace opencl {
-
 namespace detail {
 
 template <class T, class F>
@@ -57,6 +54,8 @@ T try_apply_fun(F& fun, message& msg, const T& fallback) {
 }
 
 } // namespace detail
+
+namespace opencl {
 
 // Tag classes to mark arguments received in a messages as reference or value
 /// Arguments tagged as `val` are expected as a vector (or value in case
