+++ libcaf_opencl/caf/opencl/program.hpp
@@ -23,10 +23,10 @@
 
 #include "caf/ref_counted.hpp"
 
+#include "caf/detail/raw_ptr.hpp"
+
 #include "caf/opencl/device.hpp"
 #include "caf/opencl/global.hpp"
-
-#include "caf/opencl/detail/raw_ptr.hpp"
 
 namespace caf {
 namespace opencl {
